Move NetworkTriangulation into its own source file.
Soon we'll access it from TopologyPointLocation (not just
ResolvedTopologicalNetwork).
