+++ src/api/PyReconstructionGeometries.cc
@@ -24,14 +24,13 @@
  */
 
 #include <vector>
-#include <boost/cast.hpp>
 #include <boost/foreach.hpp>
-#include <boost/operators.hpp>
 #include <boost/optional.hpp>
 
 #include "PyReconstructionGeometries.h"
 
 #include "PyInformationModel.h"
+#include "PyNetworkTriangulation.h"
 #include "PyRotationModel.h"
 #include "PythonConverterUtils.h"
 #include "PythonHashDefVisitor.h"
@@ -1878,328 +1877,6 @@ namespace GPlatesApi
 		return rigid_blocks_list;
 	}
 
-
-	/**
-	 * Information contained in Delaunay triangulation of a topological network.
-	 */
-	class NetworkTriangulation :
-			public GPlatesUtils::ReferenceCount<NetworkTriangulation>
-	{
-	public:
-
-		typedef GPlatesUtils::non_null_intrusive_ptr<NetworkTriangulation> non_null_ptr_type;
-		typedef GPlatesUtils::non_null_intrusive_ptr<const NetworkTriangulation> non_null_ptr_to_const_type;
-
-
-		class Triangle :
-				public boost::equality_comparable<Triangle>
-		{
-		public:
-			Triangle(
-					NetworkTriangulation::non_null_ptr_to_const_type triangulation,
-					const GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Face_handle &face_handle) :
-				d_triangulation(triangulation),
-				d_face_handle(face_handle)
-			{  }
-
-			bool
-			operator==(
-					const Triangle &other) const
-			{
-				return d_triangulation == other.d_triangulation &&
-						d_face_handle == other.d_face_handle;
-			}
-
-			unsigned int
-			get_vertex_index(
-					int index) const
-			{
-				if (index < 0 || index >= 3)
-				{
-					PyErr_SetString(PyExc_ValueError, "*index* should be in the range [0, 2]");
-					bp::throw_error_already_set();
-				}
-
-				return d_face_handle->vertex(index)->get_vertex_index();
-			}
-
-			bool
-			is_in_deforming_region() const
-			{
-				return d_face_handle->is_in_deforming_region();
-			}
-
-			GPlatesAppLogic::DeformationStrainRate
-			get_strain_rate() const
-			{
-				return d_face_handle->get_deformation_info().get_strain_rate();
-			}
-
-		private:
-			NetworkTriangulation::non_null_ptr_to_const_type d_triangulation;  // keep alive
-			GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Face_handle d_face_handle;
-		};
-
-		class Vertex :
-				public boost::equality_comparable<Vertex>
-		{
-		public:
-			Vertex(
-					NetworkTriangulation::non_null_ptr_to_const_type triangulation,
-					const GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Vertex_handle &vertex_handle) :
-				d_triangulation(triangulation),
-				d_vertex_handle(vertex_handle)
-			{  }
-
-			bool
-			operator==(
-					const Vertex &other) const
-			{
-				return d_triangulation == other.d_triangulation &&
-						d_vertex_handle == other.d_vertex_handle;
-			}
-
-			GPlatesMaths::PointOnSphere
-			get_position() const
-			{
-				return d_vertex_handle->get_point_on_sphere();
-			}
-
-			GPlatesMaths::Vector3D
-			get_velocity(
-					const double &velocity_delta_time,
-					GPlatesAppLogic::VelocityDeltaTime::Type velocity_delta_time_type,
-					GPlatesAppLogic::VelocityUnits::Value velocity_units,
-					const double &earth_radius_in_kms) const
-			{
-				// Velocity delta time must be positive.
-				if (velocity_delta_time <= 0)
-				{
-					PyErr_SetString(PyExc_ValueError, "Velocity delta time must be positive.");
-					bp::throw_error_already_set();
-				}
-
-				return d_vertex_handle->calc_velocity_vector(
-						velocity_delta_time,
-						velocity_delta_time_type,
-						velocity_units,
-						earth_radius_in_kms);
-			}
-
-			GPlatesAppLogic::DeformationStrainRate
-			get_strain_rate() const
-			{
-				return d_vertex_handle->get_deformation_info().get_strain_rate();
-			}
-
-		private:
-			NetworkTriangulation::non_null_ptr_to_const_type d_triangulation;  // keep alive
-			GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Vertex_handle d_vertex_handle;
-		};
-
-
-		/**
-		 * Wrapper class for functions accessing the items (triangles or vertices) of a deforming triangulation.
-		 */
-		template <class ItemType>
-		class ItemsView
-		{
-		public:
-
-			// Typedef for the item itself.
-			//
-			// Note: The item should not be none (unless something went wrong during initialisation of the triangulation).
-			typedef boost::optional<ItemType> item_type;
-			// Typedef for items sequence and iterator.
-			typedef std::vector<item_type> item_seq_type;
-			typedef typename item_seq_type::const_iterator const_iterator;
-
-			ItemsView(
-					NetworkTriangulation::non_null_ptr_type network_triangulation,
-					const item_seq_type &items) :
-				d_network_triangulation(network_triangulation),
-				d_items(items)
-			{  }
-
-			const_iterator
-			begin() const
-			{
-				return d_items.begin();
-			}
-
-			const_iterator
-			end() const
-			{
-				return d_items.end();
-			}
-
-			unsigned int
-			get_number_of_items() const
-			{
-				return d_items.size();
-			}
-
-			//
-			// Support for "__getitem__".
-			//
-			// Note: The returned item should not be none
-			//       (unless something went wrong during initialisation of the triangulation).
-			//
-			item_type
-			get_item(
-					long index) const
-			{
-				if (index < 0)
-				{
-					index += d_items.size();
-				}
-
-				if (index >= boost::numeric_cast<long>(d_items.size()) ||
-					index < 0)
-				{
-					PyErr_SetString(PyExc_IndexError, "Index out of range");
-					bp::throw_error_already_set();
-				}
-
-				return d_items[index];
-			}
-
-		private:
-			NetworkTriangulation::non_null_ptr_type d_network_triangulation;  // just to keep items reference valid
-			const item_seq_type &d_items;
-		};
-
-		typedef ItemsView<Triangle> triangles_view_type;
-		typedef ItemsView<Vertex> vertices_view_type;
-
-
-		static
-		non_null_ptr_type
-		create(
-				GPlatesAppLogic::ResolvedTopologicalNetwork::non_null_ptr_type resolved_topological_network)
-		{
-			non_null_ptr_type network_triangulation(new NetworkTriangulation(resolved_topological_network));
-
-			const GPlatesAppLogic::ResolvedTriangulation::Delaunay_2 &delaunay_triangulation_2 =
-					resolved_topological_network->get_triangulation_network().get_delaunay_2();
-
-			// Get the number of finite faces and vertices in the Delaunay triangulation.
-			const unsigned int num_faces = delaunay_triangulation_2.number_of_faces();
-			const unsigned int num_vertices = delaunay_triangulation_2.number_of_vertices();
-
-			// Resize triangle and vertex arrays to fit.
-			network_triangulation->d_triangles.resize(num_faces);
-			network_triangulation->d_vertices.resize(num_vertices);
-
-			// Iterate over the individual faces of the delaunay triangulation.
-			GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Finite_faces_iterator
-					finite_faces_2_iter = delaunay_triangulation_2.finite_faces_begin();
-			GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Finite_faces_iterator
-					finite_faces_2_end = delaunay_triangulation_2.finite_faces_end();
-			for ( ; finite_faces_2_iter != finite_faces_2_end; ++finite_faces_2_iter)
-			{
-				// Create the triangle.
-				const Triangle triangle(network_triangulation, finite_faces_2_iter);
-
-				// Index of triangle in the triangulation.
-				const unsigned int face_index = finite_faces_2_iter->get_face_index();
-				GPlatesGlobal::Assert<GPlatesGlobal::AssertionFailureException>(
-						face_index < network_triangulation->d_triangles.size(),
-						GPLATES_ASSERTION_SOURCE);
-
-				// Add the triangle.
-				network_triangulation->d_triangles[face_index] = triangle;
-			}
-
-			// Iterate over the vertices of the delaunay triangulation.
-			GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Finite_vertices_iterator
-					finite_vertices_2_iter = delaunay_triangulation_2.finite_vertices_begin();
-			GPlatesAppLogic::ResolvedTriangulation::Delaunay_2::Finite_vertices_iterator
-					finite_vertices_2_end = delaunay_triangulation_2.finite_vertices_end();
-			for ( ; finite_vertices_2_iter != finite_vertices_2_end; ++finite_vertices_2_iter)
-			{
-				// Create the vertex.
-				const Vertex vertex(network_triangulation, finite_vertices_2_iter);
-
-				// Index of vertex in the triangulation.
-				const unsigned int vertex_index = finite_vertices_2_iter->get_vertex_index();
-				GPlatesGlobal::Assert<GPlatesGlobal::AssertionFailureException>(
-						vertex_index < network_triangulation->d_vertices.size(),
-						GPLATES_ASSERTION_SOURCE);
-
-				// Add the vertex.
-				network_triangulation->d_vertices[vertex_index] = vertex;
-			}
-
-			return network_triangulation;
-		}
-
-		/**
-		 * Return the triangles in the Delaunay triangulation.
-		 *
-		 * Note: None of the returned triangles should not be none
-		 *       (unless something went wrong during initialisation of the triangulation).
-		 */
-		const std::vector<boost::optional<Triangle>> &
-		get_triangles() const
-		{
-			return d_triangles;
-		}
-
-		/**
-		 * Return the vertices in the Delaunay triangulation.
-		 *
-		 * Note: None of the returned vertices should not be none
-		 *       (unless something went wrong during initialisation of the triangulation).
-		 */
-		const std::vector<boost::optional<Vertex>> &
-		get_vertices() const
-		{
-			return d_vertices;
-		}
-
-	private:
-
-		explicit
-		NetworkTriangulation(
-				GPlatesAppLogic::ResolvedTopologicalNetwork::non_null_ptr_type resolved_topological_network) :
-			d_resolved_topological_network(resolved_topological_network)
-		{  }
-
-		// Keep the ResolvedTopologicalNetwork alive since we reference its internal Delaunay triangulation.
-		GPlatesAppLogic::ResolvedTopologicalNetwork::non_null_ptr_type d_resolved_topological_network;
-
-		std::vector<boost::optional<Triangle>> d_triangles;
-		std::vector<boost::optional<Vertex>> d_vertices;
-	};
-
-	// Need this since ".def_readonly(..., Vertex::position, ...)" does not work
-	// (because GPlatesMaths::PointGeometryOnSphere wrapped in bp::class_ instead of GPlatesMaths::PointOnSphere).
-	GPlatesMaths::PointOnSphere
-	resolved_topological_network_network_triangulation_vertex_get_position(
-			const NetworkTriangulation::Vertex &vertex)
-	{
-		return vertex.get_position();
-	}
-
-	NetworkTriangulation::triangles_view_type
-	resolved_topological_network_network_triangulation_get_triangles(
-			NetworkTriangulation::non_null_ptr_type network_triangulation)
-	{
-		return NetworkTriangulation::triangles_view_type(
-				network_triangulation,
-				network_triangulation->get_triangles());
-	}
-
-	NetworkTriangulation::vertices_view_type
-	resolved_topological_network_network_triangulation_get_vertices(
-			NetworkTriangulation::non_null_ptr_type network_triangulation)
-	{
-		return NetworkTriangulation::vertices_view_type(
-				network_triangulation,
-				network_triangulation->get_vertices());
-	}
-
 	NetworkTriangulation::non_null_ptr_type
 	resolved_topological_network_get_network_triangulation(
 			GPlatesAppLogic::ResolvedTopologicalNetwork::non_null_ptr_type resolved_topological_network)
@@ -2405,242 +2082,6 @@ namespace GPlatesApi
 void
 export_resolved_topological_network()
 {
-	{
-		//
-		// NetworkTriangulation - docstrings in reStructuredText (see http://sphinx-doc.org/rest.html).
-		//
-		bp::scope network_triangulation_wrapper_class = bp::class_<
-				GPlatesApi::NetworkTriangulation,
-				GPlatesApi::NetworkTriangulation::non_null_ptr_type,
-				boost::noncopyable>(
-						"NetworkTriangulation",
-						"The Delaunay triangulation of a :class:`resolved topological network <ResolvedTopologicalNetwork>`.\n"
-						"\n"
-						"This is the triangulation of the convex hull of the vertices obtained from the network's resolved boundary (polygon), "
-						"and any interior rigid blocks (polygons) and any interior geometries (points or lines).\n"
-						"\n"
-						".. seealso:: :ref:`pygplates_primer_network_triangulation` in the *Primer* documentation."
-						"\n"
-						".. versionadded:: 0.50\n",
-						bp::no_init)
-			.def("get_triangles",
-					&GPlatesApi::resolved_topological_network_network_triangulation_get_triangles,
-					"get_triangles\n"
-					"  Returns a read-only sequence of the triangles in this triangulation.\n"
-					"\n"
-					"  :rtype: a read-only sequence of :class:`NetworkTriangulation.Triangle`\n"
-					"\n"
-					"  The following operations for accessing the triangles in the returned read-only sequence are supported:\n"
-					"\n"
-					"  =========================== =======================================================================\n"
-					"  Operation                   Result\n"
-					"  =========================== =======================================================================\n"
-					"  ``len(seq)``                number of triangles in the triangulation\n"
-					"  ``for t in seq``            iterates over the triangles *t* in the triangulation\n"
-					"  ``seq[i]``                  the triangle in the triangulation at index *i*\n"
-					"  =========================== =======================================================================\n"
-					"\n"
-					"  The following example demonstrates some uses of the above operations:\n"
-					"  ::\n"
-					"\n"
-					"    network_triangulation = resolved_topological_network.get_network_triangulation()\n"
-					"    triangles = network_triangulation.get_triangles()\n"
-					"    vertices = network_triangulation.get_vertices()\n"
-					"    for triangle in triangles:\n"
-					"        triangle_vertex_0 = vertices[triangle.get_vertex_index(0)]\n"
-					"        triangle_vertex_1 = vertices[triangle.get_vertex_index(1)]\n"
-					"        triangle_vertex_2 = vertices[triangle.get_vertex_index(2)]\n"
-					"        triangle_strain_rate = triangle.strain_rate\n"
-					"\n"
-					"  .. note:: The returned sequence is *read-only* and cannot be modified.\n")
-			.def("get_vertices",
-					&GPlatesApi::resolved_topological_network_network_triangulation_get_vertices,
-					"get_vertices\n"
-					"  Returns a read-only sequence of the vertices in this triangulation.\n"
-					"\n"
-					"  :rtype: a read-only sequence of :class:`NetworkTriangulation.Vertex`\n"
-					"\n"
-					"  The following operations for accessing the vertices in the returned read-only sequence are supported:\n"
-					"\n"
-					"  =========================== =======================================================================\n"
-					"  Operation                   Result\n"
-					"  =========================== =======================================================================\n"
-					"  ``len(seq)``                number of vertices in the triangulation\n"
-					"  ``for v in seq``            iterates over the vertices *v* in the triangulation\n"
-					"  ``seq[i]``                  the vertex in the triangulation at index *i*\n"
-					"  =========================== =======================================================================\n"
-					"\n"
-					"  The following example demonstrates some uses of the above operations:\n"
-					"  ::\n"
-					"\n"
-					"    network_triangulation = resolved_topological_network.get_network_triangulation()\n"
-					"    vertices = network_triangulation.get_vertices()\n"
-					"    for vertex in vertices:\n"
-					"        vertex_position = vertex.position\n"
-					"        vertex_strain_rate = vertex.strain_rate\n"
-					"        vertex_velocity = vertex.get_velocity()\n"
-					"\n"
-					"  .. note:: The returned sequence is *read-only* and cannot be modified.\n")
-		;
-
-		//
-		// NetworkTriangulation.Triangle - docstrings in reStructuredText (see http://sphinx-doc.org/rest.html).
-		//
-		// A class nested within python class NetworkTriangulation (due to above 'bp::scope').
-		bp::class_<GPlatesApi::NetworkTriangulation::Triangle>(
-						"Triangle",
-						"A triangle in a :class:`network triangulation <NetworkTriangulation>`.\n"
-						"\n"
-						"Triangles are equality (``==``, ``!=``) comparable (but not hashable - cannot be used as a key in a ``dict``).\n"
-						"\n"
-						".. seealso:: :ref:`pygplates_primer_network_triangulation` in the *Primer* documentation."
-						"\n"
-						".. versionadded:: 0.50\n",
-						bp::no_init)
-			.def("get_vertex_index",
-					&GPlatesApi::NetworkTriangulation::Triangle::get_vertex_index,
-					(bp::arg("index")),
-					"get_vertex_index(index)\n"
-					"  Returns the vertex index into :meth:`NetworkTriangulation.get_vertices` of one of this triangle's three vertices.\n"
-					"\n"
-					"  :param index: the index of this triangle's vertex (in the range [0, 2])\n"
-					"  :type index: int\n"
-					"  :rtype: int\n"
-					"  :raises: ValueError if *index* is not in the range [0, 2]\n"
-					"\n"
-					"  The following example demonstrates how to access the :meth:`triangulation vertices <NetworkTriangulation.get_vertices>` "
-					"from a triangle's vertex indices:\n"
-					"  ::\n"
-					"\n"
-					"    network_triangulation = resolved_topological_network.get_network_triangulation()\n"
-					"    triangles = network_triangulation.get_triangles()\n"
-					"    vertices = network_triangulation.get_vertices()\n"
-					"    for triangle in triangles:\n"
-					"        triangle_vertex_0 = vertices[triangle.get_vertex_index(0)]\n"
-					"        triangle_vertex_1 = vertices[triangle.get_vertex_index(1)]\n"
-					"        triangle_vertex_2 = vertices[triangle.get_vertex_index(2)]\n")
-			.add_property("is_in_deforming_region",
-					&GPlatesApi::NetworkTriangulation::Triangle::is_in_deforming_region,
-					"Whether this triangle is *in* the deforming region of the network.\n"
-					"\n"
-					"  :type: bool\n"
-					"\n"
-					"  .. note:: A triangle is *in* the deforming region if its centroid is in the deforming region (where the deforming region "
-					"is defined to be *inside* the network's boundary polygon but *outside* any interior rigid block polygons).\n")
-			.add_property("strain_rate",
-					&GPlatesApi::NetworkTriangulation::Triangle::get_strain_rate,
-					"Return the constant strain rate across this triangle.\n"
-					"\n"
-					"  :type: :class:`StrainRate`\n"
-					"\n"
-					"  .. note:: This will be ``pygplates.StrainRate.zero`` if this triangle is *not* :attr:`deforming <is_in_deforming_region>`.\n")
-			// Due to the numerical tolerance in comparisons we cannot make hashable.
-			// Make unhashable, with no *equality* comparison operators (we explicitly define them)...
-			.def(GPlatesApi::NoHashDefVisitor(false, true))
-			.def(bp::self == bp::self)
-			.def(bp::self != bp::self)
-		;
-
-		//
-		// A wrapper around view access to the *triangles* of a network triangulation.
-		//
-		// We don't document this wrapper (using docstrings) since it's documented in "NetworkTriangulation".
-		bp::class_<GPlatesApi::NetworkTriangulation::triangles_view_type>(
-				// Prefix with '_' so users know it's an implementation detail (they should not be accessing it directly).
-				"_TrianglesView",
-				bp::no_init)
-			.def("__iter__",
-					bp::iterator<const GPlatesApi::NetworkTriangulation::triangles_view_type>())
-			.def("__len__",
-					&GPlatesApi::NetworkTriangulation::triangles_view_type::get_number_of_items)
-			.def("__getitem__",
-					&GPlatesApi::NetworkTriangulation::triangles_view_type::get_item)
-		;
-
-		//
-		// NetworkTriangulation.Vertex - docstrings in reStructuredText (see http://sphinx-doc.org/rest.html).
-		//
-		// A class nested within python class NetworkTriangulation (due to above 'bp::scope').
-		bp::class_<GPlatesApi::NetworkTriangulation::Vertex>(
-						"Vertex",
-						"A vertex in a :class:`network triangulation <NetworkTriangulation>`.\n"
-						"\n"
-						"Vertices are equality (``==``, ``!=``) comparable (but not hashable - cannot be used as a key in a ``dict``).\n"
-						"\n"
-						".. seealso:: :ref:`pygplates_primer_network_triangulation` in the *Primer* documentation."
-						"\n"
-						".. versionadded:: 0.50\n",
-						bp::no_init)
-			.add_property("position",
-					&GPlatesApi::resolved_topological_network_network_triangulation_vertex_get_position,
-					"Return the position of this vertex.\n"
-					"\n"
-					"  :type: :class:`PointOnSphere`\n")
-			.def("get_velocity",
-					&GPlatesApi::NetworkTriangulation::Vertex::get_velocity,
-					(bp::arg("velocity_delta_time") = 1.0,
-							bp::arg("velocity_delta_time_type") = GPlatesAppLogic::VelocityDeltaTime::T_PLUS_DELTA_T_TO_T,
-							bp::arg("velocity_units") = GPlatesAppLogic::VelocityUnits::KMS_PER_MY,
-							bp::arg("earth_radius_in_kms") = GPlatesUtils::Earth::MEAN_RADIUS_KMS),
-					"get_velocity([velocity_delta_time=1.0], [velocity_delta_time_type=pygplates.VelocityDeltaTimeType.t_plus_delta_t_to_t], "
-					"[velocity_units=pygplates.VelocityUnits.kms_per_my], [earth_radius_in_kms=pygplates.Earth.mean_radius_in_kms])\n"
-					"Returns the velocity of this vertex.\n"
-					"\n"
-					"  :param velocity_delta_time: The time delta used to calculate velocity (defaults to 1 Myr).\n"
-					"  :type velocity_delta_time: float\n"
-					"  :param velocity_delta_time_type: How the two velocity times are calculated relative to the reconstruction time. "
-					"This includes [t+dt, t], [t, t-dt] and [t+dt/2, t-dt/2]. Defaults to [t+dt, t].\n"
-					"  :type velocity_delta_time_type: *VelocityDeltaTimeType.t_plus_delta_t_to_t*, "
-					"*VelocityDeltaTimeType.t_to_t_minus_delta_t* or *VelocityDeltaTimeType.t_plus_minus_half_delta_t*\n"
-					"  :param velocity_units: whether to return velocity as *kilometres per million years* or "
-					"*centimetres per year* (defaults to *kilometres per million years*)\n"
-					"  :type velocity_units: *VelocityUnits.kms_per_my* or *VelocityUnits.cms_per_yr*\n"
-					"  :param earth_radius_in_kms: the radius of the Earth in *kilometres* (defaults to ``pygplates.Earth.mean_radius_in_kms``)\n"
-					"  :type earth_radius_in_kms: float\n"
-					"  :rtype: :class:`Vector3D`\n"
-					"  :raises: ValueError if *velocity_delta_time* is negative or zero.\n")
-			.add_property("strain_rate",
-					&GPlatesApi::NetworkTriangulation::Vertex::get_strain_rate,
-					"Return the strain rate at this vertex.\n"
-					"\n"
-					"  :type: :class:`StrainRate`\n"
-					"\n"
-					"  .. note:: This is the area-averaged strain rate of :attr:`deforming <NetworkTriangulation.Triangle.is_in_deforming_region>` "
-					"triangles incident to this vertex.\n")
-			// Due to the numerical tolerance in comparisons we cannot make hashable.
-			// Make unhashable, with no *equality* comparison operators (we explicitly define them)...
-			.def(GPlatesApi::NoHashDefVisitor(false, true))
-			.def(bp::self == bp::self)
-			.def(bp::self != bp::self)
-		;
-
-		//
-		// A wrapper around view access to the *vertices* of a network triangulation.
-		//
-		// We don't document this wrapper (using docstrings) since it's documented in "NetworkTriangulation".
-		bp::class_<GPlatesApi::NetworkTriangulation::vertices_view_type>(
-				// Prefix with '_' so users know it's an implementation detail (they should not be accessing it directly).
-				"_VerticesView",
-				bp::no_init)
-			.def("__iter__",
-					bp::iterator<const GPlatesApi::NetworkTriangulation::vertices_view_type>())
-			.def("__len__",
-					&GPlatesApi::NetworkTriangulation::vertices_view_type::get_number_of_items)
-			.def("__getitem__",
-					&GPlatesApi::NetworkTriangulation::vertices_view_type::get_item)
-		;
-	}
-
-	// Register to/from Python conversions of non_null_intrusive_ptr<> including const/non-const and boost::optional.
-	GPlatesApi::PythonConverterUtils::register_all_conversions_for_non_null_intrusive_ptr<GPlatesApi::NetworkTriangulation>();
-
-	// Enable boost::optional<NetworkTriangulation::Triangle> to be passed to and from python.
-	GPlatesApi::PythonConverterUtils::register_optional_conversion<GPlatesApi::NetworkTriangulation::Triangle>();
-
-	// Enable boost::optional<NetworkTriangulation::Vertex> to be passed to and from python.
-	GPlatesApi::PythonConverterUtils::register_optional_conversion<GPlatesApi::NetworkTriangulation::Vertex>();
-
-
 	//
 	// ResolvedTopologicalNetwork - docstrings in reStructuredText (see http://sphinx-doc.org/rest.html).
 	//
